[Xenious : 0.0.2750.0] - [Xbox360 : 0.0.1600.0] - [PowerPC : 0.0.615.0]
Decryption and Encryption is now supported in the Xbox 360 Libary.

+++ Xbox360/XEX/XeEncryptionType.cs
@@ -10,7 +10,7 @@ using System.Text;
 
 namespace Xbox360.XEX
 {
-    public enum XeEncryptionType : uint
+    public enum XeEncryptionType : UInt16
     {
         NotEncrypted = 0,
         Encrypted = 1
